[dhctl] mirror: support cse mirroring (#7017)
